% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passSim.R
\name{passSim}
\alias{passSim}
\title{passSim}
\usage{
passSim(qbdata, kappa_0, nu_0)
}
\arguments{
\item{kappa_0}{prior something}

\item{nu_0}{prior degrees of freedom}
}
\value{

}
\description{
What does this do?
}
\details{
Fill in the details
}
