\name{RIPPEN-package}
\alias{RIPPEN-package}
\alias{RIPPEN}
\docType{package}
\title{
\packageTitle{RIPPEN}
}
\description{
\packageDescription{RIPPEN}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RIPPEN}
\packageIndices{RIPPEN}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{RIPPEN}

Maintainer: \packageMaintainer{RIPPEN}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
