% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSim.R
\name{runSim}
\alias{runSim}
\title{runSim}
\usage{
runSim(passer, kappa_0 = 1, nu_0 = 3, nsim = 100)
}
\arguments{
\item{kappa_0}{prior something}

\item{nu_0}{prior degrees of freedom}
}
\value{

}
\description{
What does this do?
}
\details{
Fill in the details
}
